pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "companion-bot-core"
version = "0.1.0"
description = "Multi-User Configurable Telegram Companion Bot"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.15,<4.0",
    "asyncpg>=0.30.0",
    "redis[asyncio]>=5.2.0",
    "alembic>=1.14.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "pydantic-settings>=2.7.0",
    "structlog>=25.1.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "prometheus-client>=0.21.0",
    "cryptography>=42.0.0",
    "aiohttp>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0",
    "types-redis>=4.6.0",
    "asyncpg-stubs>=0.30.0",
    "fakeredis>=2.26.0",
]

[project.scripts]
"companion-bot-core" = "companion_bot_core.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/companion_bot_core"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "C4", "RET", "SIM", "TCH", "ANN"]
ignore = [
    "S101",    # use of assert (needed in tests)
    "ANN401",  # Any is acceptable in structlog processor signatures
]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel", "pydantic_settings.BaseSettings"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S"]
# SQLAlchemy resolves Mapped[] annotations at runtime, so stdlib types
# (datetime) used in Mapped[datetime] must remain as real imports.
"src/companion_bot_core/db/models.py" = ["TC003"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false
mypy_path = "src"
packages = ["companion_bot_core"]
exclude = ["alembic/"]

[[tool.mypy.overrides]]
module = [
    "aiogram.*",
    "asyncpg.*",
    "alembic.*",
    "structlog.*",
    "tenacity.*",
    "fakeredis.*",
    "prometheus_client.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# aiohttp ships py.typed but its internal TestClient generic parameters
# (_Request, _ApplicationNone) produce cascading type-arg errors in test files.
# Suppressing these codes for tests keeps the config stable without silencing
# meaningful production-code errors.
module = ["tests.*"]
disable_error_code = ["type-arg", "misc", "unused-ignore"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src/companion_bot_core"]
omit = ["tests/*"]
